mux: fix the volume ramp dividing integers and going NaN
rateDenom was computed as float32(n / channelCount), an integer
division. When a playing buffer had fewer samples left than the number
of channels (e.g. one stereo sample with SetBufferSize), the denominator
became 0 and the ramp rate 0/0 = NaN, so every sample of the ramp was
dropped as non-finite and the player went silent until the next
SetVolume. Divide floats instead.

+++ internal/mux/mux.go
@@ -488,7 +488,10 @@ func (p *playerImpl) readBufferAndAdd(buf []float32) int {
 	volume := float32(p.volume)
 
 	channelCount := p.mux.channelCount
-	rateDenom := float32(n / channelCount)
+	// The division must be done on floats, or a buffer with fewer samples than
+	// channels makes rateDenom 0 and the ramp rate 0/0, i.e. NaN, silencing the
+	// whole ramp.
+	rateDenom := float32(n) / float32(channelCount)
 
 	src := p.buf[:n*bitDepthInBytes]
